Print Col values as array // Check if row is win for naughts or crosses
Got a function working that prints a whole column from the game as an array. This will then be used to evaluate if the array returned is equal to a winning column for crosses or naughts. Also got a function that checks if a row is equal to a win for naughts or crosses working.